Remove global state from consultant
It's been moved to the example to show how users of the lib could use it.
