+++ src/key-parser.js
